+++ src/virtual-scroll.ts
@@ -6,16 +6,19 @@ import {
   HostListener,
   Input,
   NgModule,
+  NgZone,
   OnChanges,
   OnDestroy,
   OnInit,
   Output,
+  Renderer2,
   SimpleChanges,
   ViewChild,
 } from '@angular/core';
 
 import * as $ from 'jquery';
-var throttle = require('lodash.throttle');
+import * as tween from '@tweenjs/tween.js'
+
 export interface ChangeEvent {
   start?: number;
   end?: number;
@@ -27,9 +30,8 @@ const SCROLL_INTO_ANIM_DURATION = 400;
   selector: 'virtual-scroll,[virtualScroll]',
   exportAs: 'virtualScroll',
   template: `
-    <div class="total-padding" [style.height]="scrollHeight + 'px'"></div>
-    <div class="scrollable-content" #content [style.transform]="'translateY(' + topPadding + 'px)'"
-     [style.webkitTransform]="'translateY(' + topPadding + 'px)'">
+    <div class="total-padding" #shim></div>
+    <div class="scrollable-content" #content>
       <ng-content></ng-content>
     </div>
   `,
@@ -57,7 +59,6 @@ const SCROLL_INTO_ANIM_DURATION = 400;
   `]
 })
 export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
-
   @Input()
   items: any[] = [];
 
@@ -77,18 +78,17 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
   bufferAmount: number = 0;
 
   @Input()
-  throttleTime: number = 170;
+  scrollAnimationTime: number = 1500;
 
   private refreshHandler = () => {
     this.refresh();
   };
-  _parentScroll: Element | Window;
+  private _parentScroll: Element | Window;
   @Input()
   set parentScroll(element: Element | Window) {
     if (this._parentScroll === element) {
       return;
     }
-    this.removeParentEventHandlers(this._parentScroll);
     this._parentScroll = element;
     this.addParentEventHandlers(this._parentScroll);
   }
@@ -99,6 +99,7 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
 
   @Output()
   update: EventEmitter<any[]> = new EventEmitter<any[]>();
+  viewPortItems: any[];
 
   @Output()
   change: EventEmitter<ChangeEvent> = new EventEmitter<ChangeEvent>();
@@ -112,32 +113,42 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
   @ViewChild('content', { read: ElementRef })
   contentElementRef: ElementRef;
 
+  @ViewChild('shim', { read: ElementRef })
+  shimElementRef: ElementRef;
+
   @ContentChild('container')
   containerElementRef: ElementRef;
 
-  topPadding: number;
-  scrollHeight: number;
   previousStart: number;
   previousEnd: number;
   startupLoop: boolean = true;
-  refreshThrottled = null;
+  currentTween: any;
 
-  constructor(private element: ElementRef) {
-    this.refreshThrottled = throttle(this.refresh, this.throttleTime, {trailing:true});
-  }
+  private disposeScrollHandler: () => void | undefined;
+  private disposeResizeHandler: () => void | undefined;
 
-  @HostListener('scroll')
-  onScroll() {
-    this.refreshThrottled();
-  }
+  /** Cache of the last scroll height to prevent setting CSS when not needed. */
+  private lastScrollHeight = -1;
+
+  /** Cache of the last top padding to prevent setting CSS when not needed. */
+  private lastTopPadding = -1;
+
+  constructor(
+    private readonly element: ElementRef,
+    private readonly renderer: Renderer2,
+    private readonly zone: NgZone) { }
 
   ngOnInit() {
     this.scrollbarWidth = 0; // this.element.nativeElement.offsetWidth - this.element.nativeElement.clientWidth;
     this.scrollbarHeight = 0; // this.element.nativeElement.offsetHeight - this.element.nativeElement.clientHeight;
+
+    if (!this.parentScroll) {
+      this.addParentEventHandlers(this.element.nativeElement);
+    }
   }
 
   ngOnDestroy() {
-    this.removeParentEventHandlers(this.parentScroll);
+    this.removeParentEventHandlers();
   }
 
   ngOnChanges(changes: SimpleChanges) {
@@ -151,11 +162,13 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
   }
 
   refresh(callback: Function = undefined) {
-    requestAnimationFrame(() => {
-      this.calculateItems();
-      if (callback) {
-        callback();
-      }
+    this.zone.runOutsideAngular(() => {
+      requestAnimationFrame(() => {
+        this.calculateItems();
+        if (callback) {
+          callback();
+        }
+      });
     });
   }
 
@@ -166,6 +179,7 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
     if (index < 0 || index >= (this.items || []).length) return;
 
     let d = this.calculateDimensions();
+
     if (index >= this.previousStart && index <= this.previousEnd) {
       //can accurately scroll to a rendered item using its offsetTop
       var itemElem = document.getElementById(item.id);
@@ -188,23 +202,54 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
         this.scrollInto(item, scrollEndCallback, false);
       });
     }
+
+    let scrollTop = (Math.floor(index / d.itemsPerRow) * d.childHeight)
+      - (d.childHeight * Math.min(index, this.bufferAmount));
+
+    if (this.currentTween != undefined) this.currentTween.stop()
+    this.currentTween = new tween.Tween({ scrollTop: el.scrollTop })
+      .to({ scrollTop }, this.scrollAnimationTime)
+      .easing(tween.Easing.Quadratic.Out)
+      .onUpdate((data) => {
+        this.renderer.setProperty(el, 'scrollTop', data.scrollTop);
+        this.refresh();
+      })
+      .start();
+
+    const animate = (time?) => {
+      this.currentTween.update(time);
+      if (this.currentTween._object.scrollTop !== scrollTop) {
+        this.zone.runOutsideAngular(() => {
+          requestAnimationFrame(animate);
+        });
+      }
+    };
+
+    animate();
   }
 
   private addParentEventHandlers(parentScroll: Element | Window) {
+    this.removeParentEventHandlers();
     if (parentScroll) {
-      parentScroll.addEventListener('scroll', this.refreshHandler);
-      if (parentScroll instanceof Window) {
-        parentScroll.addEventListener('resize', this.refreshHandler);
-      }
+      this.zone.runOutsideAngular(() => {
+        this.disposeScrollHandler =
+          this.renderer.listen(parentScroll, 'scroll', this.refreshHandler);
+        if (parentScroll instanceof Window) {
+          this.disposeScrollHandler =
+            this.renderer.listen('window', 'resize', this.refreshHandler);
+        }
+      });
     }
   }
 
-  private removeParentEventHandlers(parentScroll: Element | Window) {
-    if (parentScroll) {
-      parentScroll.removeEventListener('scroll', this.refreshHandler);
-      if (parentScroll instanceof Window) {
-        parentScroll.removeEventListener('resize', this.refreshHandler);
-      }
+  private removeParentEventHandlers() {
+    if (this.disposeScrollHandler) {
+      this.disposeScrollHandler();
+      this.disposeScrollHandler = undefined;
+    }
+    if (this.disposeResizeHandler) {
+      this.disposeResizeHandler();
+      this.disposeResizeHandler = undefined;
     }
   }
 
@@ -255,9 +300,18 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
     let itemsPerRow = Math.max(1, this.countItemsPerRow());
     let itemsPerRowByCalc = Math.max(1, Math.floor(viewWidth / childWidth));
     let itemsPerCol = Math.max(1, Math.floor(viewHeight / childHeight));
-    let scrollTop = Math.max(0, el.scrollTop);
-    if (itemsPerCol === 1 && Math.floor(scrollTop / this.scrollHeight * itemCount) + itemsPerRowByCalc >= itemCount) {
+    let elScrollTop = this.parentScroll instanceof Window
+      ? (window.pageYOffset || document.documentElement.scrollTop || document.body.scrollTop || 0)
+      : el.scrollTop;
+    let scrollTop = Math.max(0, elScrollTop);
+    const scrollHeight = childHeight * itemCount / itemsPerRow;
+    if (itemsPerCol === 1 && Math.floor(scrollTop / scrollHeight * itemCount) + itemsPerRowByCalc >= itemCount) {
       itemsPerRow = itemsPerRowByCalc;
+    }
+
+    if (scrollHeight !== this.lastScrollHeight) {
+      this.renderer.setStyle(this.shimElementRef.nativeElement, 'height', `${scrollHeight}px`);
+      this.lastScrollHeight = scrollHeight;
     }
 
     return {
@@ -268,11 +322,13 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
       childHeight: childHeight,
       itemsPerRow: itemsPerRow,
       itemsPerCol: itemsPerCol,
-      itemsPerRowByCalc: itemsPerRowByCalc
+      itemsPerRowByCalc: itemsPerRowByCalc,
+      scrollHeight,
     };
   }
 
   private calculateItems() {
+    NgZone.assertNotInAngularZone();
     let el = this.parentScroll instanceof Window ? document.body : this.parentScroll || this.element.nativeElement;
     let d = this.calculateDimensions();
 
@@ -290,13 +346,16 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
 
     let items = this.items || [];
     let offsetTop = this.getElementsOffset();
-    this.scrollHeight = d.childHeight * d.itemCount / d.itemsPerRow;
-    if (el.scrollTop > this.scrollHeight) {
-      el.scrollTop = this.scrollHeight + offsetTop;
+    let elScrollTop = this.parentScroll instanceof Window
+      ? (window.pageYOffset || document.documentElement.scrollTop || document.body.scrollTop || 0)
+      : el.scrollTop;
+
+    if (elScrollTop > d.scrollHeight) {
+      elScrollTop = d.scrollHeight + offsetTop;
     }
 
-    let scrollTop = Math.max(0, el.scrollTop - offsetTop);
-    let indexByScrollTop = scrollTop / this.scrollHeight * d.itemCount / d.itemsPerRow;
+    let scrollTop = Math.max(0, elScrollTop - offsetTop);
+    let indexByScrollTop = scrollTop / d.scrollHeight * d.itemCount / d.itemsPerRow;
     let end = Math.min(d.itemCount, Math.ceil(indexByScrollTop) * d.itemsPerRow + d.itemsPerRow * (d.itemsPerCol + 1));
 
     let maxStartEnd = end;
@@ -307,7 +366,13 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
     let maxStart = Math.max(0, maxStartEnd - d.itemsPerCol * d.itemsPerRow - d.itemsPerRow);
     let start = Math.min(maxStart, Math.floor(indexByScrollTop) * d.itemsPerRow);
 
-    this.topPadding = d.childHeight * Math.ceil(start / d.itemsPerRow) - (d.childHeight * Math.min(start, this.bufferAmount));;
+    const topPadding = (items == null || items.length === 0) ? 0 : (d.childHeight * Math.ceil(start / d.itemsPerRow) - (d.childHeight * Math.min(start, this.bufferAmount)));
+
+    if (topPadding !== this.lastTopPadding) {
+      this.renderer.setStyle(this.contentElementRef.nativeElement, 'transform', `translateY(${topPadding}px)`);
+      this.renderer.setStyle(this.contentElementRef.nativeElement, 'webkitTransform', `translateY(${topPadding}px)`);
+      this.lastTopPadding = topPadding;
+    }
 
     start = !isNaN(start) ? start : -1;
     end = !isNaN(end) ? end : -1;
@@ -317,27 +382,30 @@ export class VirtualScrollComponent implements OnInit, OnChanges, OnDestroy {
     end = Math.min(items.length, end);
     if (start !== this.previousStart || end !== this.previousEnd) {
 
-      // update the scroll list
-      this.update.emit(items.slice(start, end));
+      this.zone.run(() => {
+        // update the scroll list
+        this.viewPortItems = items.slice(start, end);
+        this.update.emit(this.viewPortItems);
 
-      // emit 'start' event
-      if (start !== this.previousStart && this.startupLoop === false) {
-        this.start.emit({ start, end });
-      }
+        // emit 'start' event
+        if (start !== this.previousStart && this.startupLoop === false) {
+          this.start.emit({ start, end });
+        }
 
-      // emit 'end' event
-      if (end !== this.previousEnd && this.startupLoop === false) {
-        this.end.emit({ start, end });
-      }
+        // emit 'end' event
+        if (end !== this.previousEnd && this.startupLoop === false) {
+          this.end.emit({ start, end });
+        }
 
-      this.previousStart = start;
-      this.previousEnd = end;
+        this.previousStart = start;
+        this.previousEnd = end;
 
-      if (this.startupLoop === true) {
-        this.refresh();
-      } else {
-        this.change.emit({ start, end });
-      }
+        if (this.startupLoop === true) {
+          this.refresh();
+        } else {
+          this.change.emit({ start, end });
+        }
+      });
 
     } else if (this.startupLoop === true) {
       this.startupLoop = false;